+++ include/wfa/art_bootstrap_execution_fixture.hpp
@@ -29,6 +29,7 @@ struct NativeArtBootstrapExecutionFixtureReport {
   std::string selected_activity_class_descriptor;
   std::string application_bootstrap_command;
   std::string activity_bootstrap_command;
+  std::string art_runtime_probe_source;
   bool art_runtime_detected = false;
   bool safe_runtime_probe_available = false;
   bool runtime_class_resolution_succeeded = false;